+++ app/Http/Livewire/ConfiguratorForm.php
@@ -9,6 +9,7 @@ use Symfony\Component\Yaml\Yaml;
 
 /**
  * Class ConfiguratorForm
+ *
  * @package App\Http\Livewire
  */
 class ConfiguratorForm extends Component
@@ -21,6 +22,8 @@ class ConfiguratorForm extends Component
     public $manualTrigger;
     public $mysqlService;
     public $mysqlDatabase;
+    public $mysqlPasswordType; // 'skip
+    public $mysqlPassword; // password
     public $mysqlVersion;
     public $mysqlDatabaseName;
     public $mysqlDatabasePort;
@@ -30,7 +33,12 @@ class ConfiguratorForm extends Component
     public $stepNodejsVersion; // 12.x
     public $stepCachePackages; //true
     public $stepCacheVendors; //true
+    public $stepCacheNpmModules; // true
     public $stepFixStoragePermissions; //true
+    public $stepRunMigrations; // true
+    public $stepExecutePhpunit; //true
+    public $stepExecuteCodeSniffer; //false
+    public $stepExecuteStaticAnalysis; // false
     public $stepDusk; // false
 
     public $result;
@@ -38,39 +46,56 @@ class ConfiguratorForm extends Component
 
     public function mount()
     {
-        $this->name="Test Laravel Github action";
+        $this->name = "Test Laravel Github action";
         $this->onPush = true;
         $this->onPushBranches = ["main", "develop", "features/**"];
-        $this->onPullrequest= false;
+        $this->onPullrequest = false;
         $this->onPullrequestBranches = ["main", "develop"];
         $this->manualTrigger = false;
         $this->mysqlService = true;
-        $this->mysqlDatabase="mysql";
-        $this->mysqlVersion="5.7";
+        $this->mysqlDatabase = "mysql";
+        $this->mysqlPasswordType = "skip";
+        $this->mysqlPassword = "DB_PASSWORD";
+
+        $this->mysqlVersion = "5.7";
         $this->mysqlDatabaseName = "db_test_laravel";
         $this->mysqlDatabasePort = 33306;
-        $this->stepEnvTemplateFile= ".env.example";
-        $this->stepPhpVersions= ["8.0", "7.4"];
+        $this->stepEnvTemplateFile = ".env.example";
+        $this->stepPhpVersions = ["8.0", "7.4"];
         $this->stepNodejs = false;
-        $this->stepNodejsVersion ="12.x";
-        $this->stepCachePackages =true;
+        $this->stepNodejsVersion = "14.x";
+        $this->stepCachePackages = true;
         $this->stepCacheVendors = true;
+        $this->stepCacheNpmModules  = true;
         $this->stepFixStoragePermissions = true;
+        $this->stepRunMigrations = true;
+        $this->stepExecutePhpunit = true;
+        $this->stepExecuteCodeSniffer = false;
+        $this->stepExecuteStaticAnalysis = false;
         $this->stepDusk = false;
         $this->result = "";
         $this->errorGeneration = "";
     }
 
-    private static function split($somethingToSplit, $splitChars=",")
+    private static function split($somethingToSplit, $splitChars = ",")
     {
-        if (\is_string( $somethingToSplit)) {
+        if (\is_string($somethingToSplit)) {
             return array_map('trim', explode($splitChars, $somethingToSplit));
         }
         return $somethingToSplit;
     }
 
-    private static function arrayToString($array) {
-        return "[ " . implode(",", array_map(function ($str) { return "'$str'"; }, $array)) . " ]";
+    private static function arrayToString($array)
+    {
+        return "[ " . implode(
+            ",",
+            array_map(
+                function ($str) {
+                    return "'$str'";
+                },
+                $array
+            )
+        ) . " ]";
     }
 
     private function compactThis(...$args)
@@ -95,6 +120,8 @@ class ConfiguratorForm extends Component
             "mysqlVersion",
             "mysqlDatabaseName",
             "mysqlDatabasePort",
+            "mysqlPassword",
+            "mysqlPasswordType",
             "name",
             "on_push",
             "on_push_branches",
@@ -107,10 +134,18 @@ class ConfiguratorForm extends Component
             "stepNodejsVersion",
             "stepCachePackages",
             "stepCacheVendors",
+            "stepCacheNpmModules",
             "stepFixStoragePermissions",
+            "stepRunMigrations",
+            "stepExecutePhpunit",
+            "stepExecuteCodeSniffer",
+            "stepExecuteStaticAnalysis",
             "stepDusk"
         );
         $data["stepPhpVersionsString"] = self::arrayToString($this->stepPhpVersions);
+        $data["on_pullrequest_branches"] = self::split($this->onPullrequestBranches);
+        $data["on_push_branches"] = self::split($this->onPushBranches);
+
         $stringResult = view('action_yaml', $data)->render();
 
         try {
@@ -119,12 +154,8 @@ class ConfiguratorForm extends Component
             $this->result = $stringResult;
         } catch (ParseException $e) {
             $this->errorGeneration = $e->getMessage();
+            $this->result = $stringResult;
         }
-
-
-
-
-
     }
 
     public function render()